fix: remove chat circular dep
