+++ frontend/src/lib/components/copilot/chat/shared.ts
@@ -7,6 +7,10 @@ import { get } from 'svelte/store'
 import type { ContextElement } from './context'
 import { workspaceStore } from '$lib/stores'
 import type { ExtendedOpenFlow } from '$lib/components/flows/types'
+import type { FunctionParameters } from 'openai/resources/shared.mjs'
+import { zodToJsonSchema } from 'zod-to-json-schema'
+import { z } from 'zod'
+import { ScriptService } from '$lib/gen'
 
 type BaseDisplayMessage = {
 	content: string
@@ -112,3 +116,77 @@ export interface Tool<T> {
 export interface ToolCallbacks {
 	setToolStatus: (id: string, content: string) => void
 }
+
+export function createToolDef(
+	zodSchema: z.ZodSchema,
+	name: string,
+	description: string
+): ChatCompletionTool {
+	const schema = zodToJsonSchema(zodSchema, {
+		name,
+		target: 'openAi'
+	})
+	let parameters = schema.definitions![name] as FunctionParameters
+	parameters = {
+		...parameters,
+		required: parameters.required ?? []
+	}
+
+	return {
+		type: 'function',
+		function: {
+			strict: true,
+			name,
+			description,
+			parameters
+		}
+	}
+}
+
+const searchHubScriptsSchema = z.object({
+	query: z
+		.string()
+		.describe('The query to search for, e.g. send email, list stripe invoices, etc..')
+})
+
+const searchHubScriptsToolDef = createToolDef(
+	searchHubScriptsSchema,
+	'search_hub_scripts',
+	'Search for scripts in the hub'
+)
+
+export const createSearchHubScriptsTool = (withContent: boolean = false) => ({
+	def: searchHubScriptsToolDef,
+	fn: async ({ args, toolId, toolCallbacks }) => {
+		toolCallbacks.setToolStatus(
+			toolId,
+			'Searching for hub scripts related to "' + args.query + '"...'
+		)
+		const parsedArgs = searchHubScriptsSchema.parse(args)
+		const scripts = await ScriptService.queryHubScripts({
+			text: parsedArgs.query,
+			kind: 'script'
+		})
+		toolCallbacks.setToolStatus(
+			toolId,
+			'Found ' + scripts.length + ' scripts in the hub related to "' + args.query + '"'
+		)
+		// if withContent, fetch scripts with their content, limit to 3 results
+		const results = await Promise.all(
+			scripts.slice(0, withContent ? 3 : undefined).map(async (s) => {
+				let content = ''
+				if (withContent) {
+					content = await ScriptService.getHubScriptContentByPath({
+						path: `hub/${s.version_id}/${s.app}/${s.summary.toLowerCase().replaceAll(/\s+/g, '_')}`
+					})
+				}
+				return {
+					path: `hub/${s.version_id}/${s.app}/${s.summary.toLowerCase().replaceAll(/\s+/g, '_')}`,
+					summary: s.summary,
+					...(withContent ? { content } : {})
+				}
+			})
+		)
+		return JSON.stringify(results)
+	}
+})
